Change タスク処理をワーカへ委任

+++ main/task/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Column, ForeignKey, Integer, String, DateTime, JSON
+from sqlalchemy import Boolean, Column, ForeignKey, Integer, String, DateTime, JSON, Float
 from sqlalchemy.orm import relationship
 from mixin.database import Base, Engine
 
@@ -7,7 +7,7 @@ class TaskModel(Base):
     __tablename__ = "queue"
     uuid = Column(String, primary_key=True, index=True)
     post_time = Column(DateTime)
-    run_time = Column(Integer)
+    run_time = Column(Float)
     user_id = Column(Integer(), ForeignKey('users.user_id', onupdate='CASCADE', ondelete='CASCADE'))
     status = Column(String)
     resource = Column(String)
